+++ tsup.config.ts
@@ -3,7 +3,9 @@ import { defineConfig } from "tsup";
 export default defineConfig({
   entry: ["src/index.ts"],
   format: ["cjs", "esm"],
-  dts: true,
+  dts: {
+    banner: 'import "zustand/middleware";',
+  },
   splitting: false,
   sourcemap: true,
   clean: true,